+++ changelog.md
@@ -1,7 +1,8 @@
 InsaneLib's Changelog
 
 ## 1.8.1
-* Added ILItemTier
+* Added `ILItemTier`
+* Added `Feature#getConfigOption` (and renamed `Feature#setConfig` to `Feature#setConfigOption`)
 * Fixed nested Blacklist and MinMax configs not popping the path correctly
 
 ## 1.8.0